PEThorn restricts parameters to the validated range in U.S.A.
Thornthwaite equations break when they are used within their
validate range in U.S.A., or in cool boreal and arctic climate or
in hot tropics. Too high PET is predicted to the arctic, and too
low to the tropics. Now we keep the range of "heat index" in the
validate U.S. limits, and force the index to the limit in the north
and in the south.