Add 'twist' to implement simpler estimation of argument 'A'
PEThorn tries to use a 3rd degree polynomial to produce a
straight line, and almost succeeds -- up to limits of validity. Now
we have a more exact and simpler method that is not limited by the
range of Ival. More over, this is also more exact in its goal. This
does not mean the predicted PET would be more correct, though --
quite the contrary.