as.vector is no longer exported from raster (although still documented)
Is looks badly that as.vector was removed from the raster package
to an unkonwn base. It may be that base::as.vector will work, and
to try this the raster::as.vector import was removed. However,
I cannot test this now since proj4 library is broken in
ubuntu and CRS do not work, and examples end with error when
trying to set the projection...